Add files via upload
In this model the user inputs a book description then the model recommends a publisher that has published books similar to the description given.
The objective is to find a publisher for an author that would most likely publish their book, based on that publishers past history. 
The dataset, available on Kaggle, includes data collected from Goodreads.
The model was trained using the tfidf approach.
The initial dataset was approximately 2,000,000 rows. However this was truncated due to hardware limitations.